+++ Realm/Tests/ObjectTests.m
@@ -776,4 +776,34 @@ RLM_ARRAY_TYPE(CycleObject)
     OSSpinLockLock(&spinlock);
 }
 
+- (void)testIsDeleted {
+    StringObject *obj1 = [[StringObject alloc] initWithObject:@[@"a"]];
+    XCTAssertEqual(obj1.deletedFromRealm, NO);
+
+    RLMRealm *realm = [self realmWithTestPath];
+    [realm beginWriteTransaction];
+    [realm addObject:obj1];
+    StringObject *obj2 = [StringObject createInRealm:realm withObject:@[@"b"]];
+
+    XCTAssertEqual([obj1 isDeletedFromRealm], NO);
+    XCTAssertEqual(obj2.deletedFromRealm, NO);
+
+    [realm commitWriteTransaction];
+
+    // delete
+    [realm beginWriteTransaction];
+    [realm deleteObject:obj1];
+    [realm deleteObject:obj2];
+
+    XCTAssertEqual(obj1.deletedFromRealm, YES);
+    XCTAssertEqual(obj2.deletedFromRealm, YES);
+
+    XCTAssertThrows([realm addObject:obj1], @"Adding deleted object should throw");
+    
+    [realm commitWriteTransaction];
+
+    XCTAssertEqual(obj1.deletedFromRealm, YES);
+    XCTAssertNil(obj1.realm, @"Realm should be nil after deletion");
+}
+
 @end